test(eve): consolidate channel continuation e2e
